core: move template specilization to the right place
