+++ CMakeLists.txt
@@ -120,7 +120,6 @@ SET(${PROJECT_NAME}_HEADERS
   include/eigenpy/quaternion.hpp
   include/eigenpy/stride.hpp
   include/eigenpy/ref.hpp
-  include/eigenpy/details/rvalue_from_python_data.hpp
   include/eigenpy/version.hpp
 )
