+++ include/eigenpy/details.hpp
@@ -6,7 +6,6 @@
 #ifndef __eigenpy_details_hpp__
 #define __eigenpy_details_hpp__
 
-#include "eigenpy/details/rvalue_from_python_data.hpp"
 #include "eigenpy/fwd.hpp"
 #include "eigenpy/eigenpy.hpp"
